+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUIRelayMK3.java
@@ -27,7 +27,7 @@ public class GUIRelayMK3 extends GuiContainer
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
 		this.fontRendererObj.drawString("Relay MK3", 38, 6, 4210752);
-		this.fontRendererObj.drawString(Integer.toString(tile.displayEmc), 125, 39, 4210752);
+		this.fontRendererObj.drawString(Double.toString(tile.displayEmc), 125, 39, 4210752);
 	}
 
 	@Override
@@ -42,15 +42,15 @@ public class GUIRelayMK3 extends GuiContainer
 		this.drawTexturedModalRect(x, y, 0, 0, xSize, ySize);
 		
 		//Emc bar progress
-		int progress = tile.getEmcScaled(102);
-		this.drawTexturedModalRect(x + 105, y + 6, 30, 195, progress, 10);
+		double progress = tile.getEmcScaled(102);
+		this.drawTexturedModalRect(x + 105, y + 6, 30, 195, (int)progress, 10);
 		
 		//Klein start bar progress. Max is 30.
 		progress = tile.getKleinEmcScaled(30);
-		this.drawTexturedModalRect(x + 153, y + 82, 0, 195, progress, 10);
+		this.drawTexturedModalRect(x + 153, y + 82, 0, 195, (int)progress, 10);
 				
 		//Burn Slot bar progress. Max is 30.
 		progress = tile.getRawEmcScaled(30);
-		drawTexturedModalRect(x + 101, y + 82, 0, 195, progress, 10);
+		drawTexturedModalRect(x + 101, y + 82, 0, 195, (int)progress, 10);
 	}	
 }
